requirements.txt
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pymongo>=4.6.0
motor>=3.3.0
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
pydantic>=2.5.0
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
requests>=2.31.0
reportlab>=4.0.0
